fix command delay bug
